// src/interfaces/employees.ts
export interface IEmployeeEntity {
  EmployeeId: string;
  JobKeyId: string;
  JobKeyName: string;
  PositionId: string;
  PositionName: string;
  DepartmentId: string;
  DepartmentName: string;
  DivisionId: string;
  DivisionName: string;
  PersonnelAreaId: string;
  PersonnelAreaName: string;
  ManagerId: string;
  Fullname: string;
  Nickname: string;
  Email: string;
  Telephone: string;
  BeginDate: string;
  EndDate: string;
  Status: string;
  CreatedDate: string;
  ModifiedDate: string;
  Operated: string;
  FullnameEN: string;
  Secrets: string;
}